refactor the next_job_id reading and writing to use a mutex
The previous implementation with two different atomic operations left a
window between them in which the same ID could be read twice by
different goroutines.

I also removed the test that was not actually testing the worker pool,
but the Go library we already trust to be correct and covered by
regression tests.